@@ -12,47 +12,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.1.10",
-    "@angular/common": "6.1.10",
-    "@angular/compiler": "6.1.10",
-    "@angular/core": "6.1.10",
-    "@angular/forms": "6.1.10",
-    "@angular/http": "6.1.10",
-    "@angular/platform-browser": "6.1.10",
-    "@angular/platform-browser-dynamic": "6.1.10",
-    "@angular/platform-server": "6.1.10",
-    "@angular/router": "6.1.10",
-    "@nguniversal/module-map-ngfactory-loader": "6.0.0",
-    "core-js": "^2.5.4",
-    "rxjs": "6.0.0",
-    "zone.js": "^0.8.26",
-    "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.3.1",
-    "jquery": "3.3.1",
-    "popper.js": "^1.14.3"
+    "@angular/animations": "~8.0.1",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "~8.0.1",
+    "@angular/compiler": "~8.0.1",
+    "@angular/core": "~8.0.1",
+    "@angular/flex-layout": "^8.0.0-beta.26",
+    "@angular/forms": "~8.0.1",
+    "@angular/platform-browser": "^8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.1",
+    "@angular/platform-server": "^8.0.3",
+    "@angular/router": "~8.0.1",
+    "@nguniversal/module-map-ngfactory-loader": "^7.1.1",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.0",
-    "@angular/cli": "~6.0.0",
-    "@angular/compiler-cli": "6.1.10",
-    "@angular/language-service": "^6.0.0",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "~8.0.3",
+    "@angular/compiler-cli": "~8.0.1",
+    "@angular/language-service": "~8.0.1",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "typescript": "~2.7.2"
-  },
-  "optionalDependencies": {
-    "node-sass": "^4.9.3",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1"
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   }
 }
